# apps/api/openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediaCompressor API",
    "version": "0.1.0",
    "description": "REST API for MediaCompressor — image/video compression service"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc_session"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["queued", "processing", "succeeded", "failed", "canceled", "expired"]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": ["image", "video"]
            },
            "in": "query",
            "name": "kind",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/jobs/{id}/events": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/internal/uploads/hooks/pre-create": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Type": {
                    "type": "string"
                  },
                  "Event": {
                    "type": "object",
                    "properties": {
                      "Upload": {
                        "type": "object",
                        "properties": {
                          "ID": {
                            "type": "string"
                          },
                          "Size": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "MetaData": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            },
                            "default": {}
                          }
                        },
                        "required": ["Size"],
                        "additionalProperties": false
                      }
                    },
                    "required": ["Upload"],
                    "additionalProperties": false
                  }
                },
                "required": ["Event"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/internal/uploads/hooks/post-finish": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Type": {
                    "type": "string"
                  },
                  "Event": {
                    "type": "object",
                    "properties": {
                      "Upload": {
                        "type": "object",
                        "properties": {
                          "ID": {
                            "type": "string",
                            "minLength": 1
                          },
                          "Size": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "Storage": {
                            "type": "object",
                            "properties": {
                              "Type": {
                                "type": "string"
                              },
                              "Path": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": false
                          },
                          "MetaData": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            },
                            "default": {}
                          }
                        },
                        "required": ["ID", "Size"],
                        "additionalProperties": false
                      }
                    },
                    "required": ["Upload"],
                    "additionalProperties": false
                  }
                },
                "required": ["Event"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "processing",
                "succeeded",
                "failed",
                "canceled",
                "expired",
                "uploading"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "fragment",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "fragment",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cancelflash",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}